+++ client/css/components/manage_session.scss
@@ -2,5 +2,4 @@
 .ql-manage-session {
 
 
-
 }
