+++ imports/ui/pages/student/student_dashboard.jsx
@@ -6,9 +6,9 @@
 import React, { Component } from 'react'
 import { createContainer } from 'meteor/react-meteor-data'
 
-import { Courses } from '../../api/courses.js'
-import { EnrollCourseModal } from '../modals/EnrollCourseModal'
-import { CourseListItem } from '../CourseListItem'
+import { Courses } from '../../../api/courses.js'
+import { EnrollCourseModal } from '../../modals/EnrollCourseModal'
+import { CourseListItem } from '../../CourseListItem'
 
 class _StudentDashboard extends Component {
   constructor (props) {
